Issue 25 submit score bug (#27)
* Log score

* Cast to long after going from C to Java

* Removed logging

+++ gpgs/src/java/com/defold/gpgs/GpgsJNI.java
@@ -611,9 +611,9 @@ public class GpgsJNI {
         return true;
     }
 
-    public void submitScore(String leaderboardId, long score) {
+    public void submitScore(String leaderboardId, double score) {
         if(initLeaderboards()) {
-            mLeaderboardsClient.submitScore(leaderboardId, score);
+            mLeaderboardsClient.submitScore(leaderboardId, (long)score);
         }
     }
